requirements.txt
streamlit 
langchain 
langchain-community
together 
openai 
python-dotenv 
earthengine-api 
geemap 
leafmap
setuptools